Refactor: Split entities into individual files
